moved utility to new package
