+++ src/main/java/nl/tudelft/model/pickups/utility/Utility.java
@@ -1,4 +1,4 @@
-package nl.tudelft.model.pickups;
+package nl.tudelft.model.pickups.utility;
 
 
 public class Utility {
