+++ src/test/java/nl/tudelft/model/LevelTest.java
@@ -14,7 +14,7 @@ import java.util.LinkedList;
 
 import nl.tudelft.model.bubble.Bubble;
 import nl.tudelft.model.pickups.Pickup;
-import nl.tudelft.model.pickups.Utility;
+import nl.tudelft.model.pickups.utility.Utility;
 import nl.tudelft.model.pickups.weapon.Projectile;
 import nl.tudelft.semgroup4.Modifiable;
